requirements.txt
annotated-types==0.6.0
asttokens==2.4.1
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
iniconfig==2.0.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
jupyter_client==8.5.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pexpect==4.8.0
platformdirs==3.11.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.16.1
pytest==7.4.3
python-dateutil==2.8.2
python-frontmatter==1.0.0
PyYAML==6.0.1
pyzmq==25.1.1
ruff==0.1.9
six==1.16.0
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.9.0
wcwidth==0.2.9
